+++ src/model/transformation/MoveT.java
@@ -32,8 +32,8 @@ public class MoveT extends AbstractTransformation {
     if (frame < 0) {
       throw new IllegalArgumentException("Frame cannot be negative.");
     }
-    int newX = this.getValueAtFrame(frame, this.startPos.getX(), this.endPos.getX());
-    int newY = this.getValueAtFrame(frame, this.startPos.getY(), this.endPos.getY());
+    float newX = this.getValueAtFrame(frame, this.startPos.getX(), this.endPos.getX());
+    float newY = this.getValueAtFrame(frame, this.startPos.getY(), this.endPos.getY());
     this.shape.setPosition(newX, newY);
 
     return this.shape;
